config: enable zstd package
trace-cmd lab currently fails with the remote setup: generated trace.dat on
host is 0 byte, and kernelshark confirms that there is no data in it when
opened.
On host side, there is an error appearing for failed sessions:

reading client -110 (unknown error)

This error is in fact a timeout during the initial exchanges between target
and host over tcp. It appears that the target uses more than 5s (the
configured timeout) at a specific step. Adding some logs show that this
step is the one during which trace-cmd gathers system data (cpu info,
kallsyms, etc). Running perf on this command shows that 2/3 of the time
used during those steps is passed in longest_match, a function from zlib.
Checking back on previous versions of the labs (trace-cmd v2.9.7), we see
that data was not compressed, which explains why we did not have the issue
(checking generating trace.dat files before/after the upgrade validate this
too )
Now, zlib was known to be slow when it was introduced, and was quickly
replaced as the default compression method by zstd (see [1]), which comes
with way better performance regarding compression time, and slightly better
performance regarding compression size.
The final twist is that our current buildroot configuration enables zlib
but not zstd: trace-cmd is then build with zlib as default and sole
compression method (there is also a "no compression" mode available)

Enable zstd in our configuration so that trace-cmd can work again for our
system-wide profiling lab

[1] https://lore.kernel.org/linux-trace-devel/20220219230151.56282-1-sebastian@breakpoint.cc/

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>